Move the ViewRenderService to its own file
